pyproject.toml
[project]
name = "code_execution_deep_agent"
version = "1.0.0"
description = "Jan Franco's implementation of a code execution deep research agent using LangGraph"
authors = [
    { name = "Jan Franco Glanc Gomez", email = "jglancgomez@gmail.com" },
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.2.5",
    "deepagents-cli>=0.0.9",
    "langchain>=1.0.5",
    "langchain-anthropic>=1.0.2",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.0",
    "langsmith>=0.4.41",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.7.12",
    "pandas>=2.0.0",
    "pypdf>=3.0.0",
    "pyyaml>=6.0.0",
    "reportlab>=4.0.0",
    "bashlex>=0.18",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
